# contract/openapi.before.json
{
  "components": {
    "schemas": {
...
      "AgriTrackSyncIn": {
        "description": "One AgriTrack sync push: a farmer and their farms.",
        "properties": {
          "boundaries": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Boundaries"
          },
          "farm": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Farm"
          },
          "farmId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Farmid"
          },
          "farmer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/_FarmerSyncIn"
              },
              {
                "type": "null"
              }
            ]
          },
          "farmerId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Farmerid"
          },
          "farms": {
            "items": {
              "$ref": "#/components/schemas/_FarmSyncIn"
            },
            "title": "Farms",
            "type": "array"
          },
          "fields": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/_FlatFieldIn"
            },
            "title": "Fields",
            "type": "array"
          },
          "subPlots": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/_FlatSubPlotIn"
            },
            "title": "Subplots",
            "type": "array"
          }
        },
        "required": [
          "farms"
        ],
        "title": "AgriTrackSyncIn",
        "type": "object"
      },
...
      "FarmIngestReport": {
        "description": "Full ingestion outcome for one farm.",
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "canonical_farm_id": {
            "title": "Canonical Farm Id",
            "type": "string"
          },
          "farm_id": {
            "title": "Farm Id",
            "type": "string"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/FieldIngestReport"
            },
            "title": "Fields",
            "type": "array"
          },
          "working_crs": {
            "title": "Working Crs",
            "type": "string"
          }
        },
        "required": [
          "canonical_farm_id",
          "farm_id",
          "action",
          "working_crs"
        ],
        "title": "FarmIngestReport",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SatelliteInterpretation": {
        "description": "The optional interpretation block (gateway spec §9). `summary` is the canonical primary\nnarrative; `notes` is kept as a backward-compat alias for existing AgriTrack consumers.\n`anomalies` is preserved here (also present in `SatelliteMetrics` per §11) for backward\ncompat. `zones` carries optional stress-zone breakdowns.",
        "properties": {
          "anomalies": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Anomalies",
            "type": "array"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "stress_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stress Level"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "zones": {
            "default": [],
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Zones",
            "type": "array"
          }
        },
        "title": "SatelliteInterpretation",
        "type": "object"
      },
      "SatelliteMetrics": {
        "description": "The flattened per-record metrics AgriTrack expects (ADR 0006). `savi_mean`/`ndre_mean` were\nadded 2026-06-19 (ADR 0006 §3 amendment): both indices were already computed and stored, but\nhad no wire field here, so they were dropped at this adapter and never reached AgriTrack.\n`anomalies` moved here from `SatelliteInterpretation` per gateway spec §11 (2026-06-21).",
        "properties": {
          "anomalies": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Anomalies",
            "type": "array"
          },
          "classification": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classification"
          },
          "cloud_cover_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cloud Cover Pct"
          },
          "evi_mean": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evi Mean"
          },
          "health_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Health Score"
          },
          "ndre_mean": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ndre Mean"
          },
          "ndvi_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ndvi Max"
          },
          "ndvi_mean": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ndvi Mean"
          },
          "ndvi_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ndvi Min"
          },
          "ndwi_mean": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ndwi Mean"
          },
          "savi_mean": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Savi Mean"
          }
        },
        "title": "SatelliteMetrics",
        "type": "object"
      },
      "SatelliteResult": {
        "description": "One AgriTrack `/integrations/satellite/results` record.\n\nFor `scope in (\"farm\", \"field\")`: flat record with top-level `extId`, `metrics`,\n`interpretation`.  For `scope == \"sub_plot\"`: grouped record with a `subPlots` array; the\nflat fields are absent (excluded from the JSON body). Field names are the contract's camelCase\nwire keys.",
        "properties": {
          "analysisDate": {
            "title": "Analysisdate",
            "type": "string"
          },
          "extId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extid"
          },
          "farmId": {
            "title": "Farmid",
            "type": "integer"
          },
          "fieldId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fieldid"
          },
          "interpretation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SatelliteInterpretation"
              },
              {
                "type": "null"
              }
            ]
          },
          "metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SatelliteMetrics"
              },
              {
                "type": "null"
              }
            ]
          },
          "outputs": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outputs"
          },
          "scope": {
            "title": "Scope",
            "type": "string"
          },
          "sourceSystem": {
            "default": "satellite",
            "title": "Sourcesystem",
            "type": "string"
          },
          "subPlotId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subplotid"
          },
          "subPlots": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SubPlotEntry"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subplots"
          }
        },
        "required": [
          "farmId",
          "scope",
          "analysisDate"
        ],
        "title": "SatelliteResult",
        "type": "object"
      },
...
      "SubPlotEntry": {
        "description": "One sub-plot within a grouped `scope: sub_plot` record (gateway spec 2026-06-18).",
        "properties": {
          "extId": {
            "title": "Extid",
            "type": "string"
          },
          "metrics": {
            "$ref": "#/components/schemas/SatelliteMetrics"
          },
          "subPlotId": {
            "title": "Subplotid",
            "type": "integer"
          }
        },
        "required": [
          "subPlotId",
          "extId",
          "metrics"
        ],
        "title": "SubPlotEntry",
        "type": "object"
      },
      "SyncResponse": {
        "description": "Per-sync outcome: how many farms were ingested and each farm's ingestion report.",
        "properties": {
          "farms": {
            "items": {
              "$ref": "#/components/schemas/FarmIngestReport"
            },
            "title": "Farms",
            "type": "array"
          },
          "synced": {
            "title": "Synced",
            "type": "integer"
          }
        },
        "required": [
          "synced",
          "farms"
        ],
        "title": "SyncResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "_FarmSyncIn": {
        "properties": {
          "boundary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Boundary"
          },
          "farm_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              }
            ],
            "title": "Farm Id"
          },
          "fields": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/_FieldSyncIn"
            },
            "title": "Fields",
            "type": "array"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "required": [
          "farm_id"
        ],
        "title": "_FarmSyncIn",
        "type": "object"
      },
      "_FarmerSyncIn": {
        "properties": {
          "agritrack_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agritrack Id"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "_FarmerSyncIn",
        "type": "object"
      },
...
      "_FlatFieldIn": {
        "description": "Gateway §6.2 canonical flat-format field (camelCase keys).",
        "properties": {
          "area_ha": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Ha"
          },
          "boundary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Boundary"
          },
          "crop": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crop"
          },
          "fieldId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              }
            ],
            "title": "Fieldid"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "required": [
          "fieldId"
        ],
        "title": "_FlatFieldIn",
        "type": "object"
      },
      "_FlatSubPlotIn": {
        "description": "Gateway §6.2 canonical flat-format sub-plot; fieldId links it to its parent field.",
        "properties": {
          "boundary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Boundary"
          },
          "crop": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crop"
          },
          "fieldId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              }
            ],
            "title": "Fieldid"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "subPlotId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              }
            ],
            "title": "Subplotid"
          }
        },
        "required": [
          "subPlotId",
          "fieldId"
        ],
        "title": "_FlatSubPlotIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "remote-sense API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/mobile/data": {
      "get": {
        "description": "Pull a farm's stored satellite results in the AgriTrack contract shape (ADR 0006 section 4);\nAgriTrack triggers this per farm when a push could not be delivered. Geometry is never returned\n(invariant 6).",
        "operationId": "mobile_data_api_v1_mobile_data_get",
        "parameters": [
          {
            "in": "query",
            "name": "farm_id",
            "required": true,
            "schema": {
              "title": "Farm Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SatelliteResult"
                  },
                  "title": "Response Mobile Data Api V1 Mobile Data Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mobile Data",
        "tags": [
          "agritrack"
        ]
      }
    },
    "/api/v1/mobile/sync": {
      "post": {
        "description": "Ingest an AgriTrack sync (farmer -> farms -> fields -> sub-plots). Idempotent: re-sending the\nsame payload changes nothing (the ingestion layer dedupes on the canonical ids). Bad geometry is\nrejected with 422.",
        "operationId": "mobile_sync_api_v1_mobile_sync_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgriTrackSyncIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mobile Sync",
        "tags": [
          "agritrack"
        ]
      }
    },
...
  }
}